[setup] Refine install scripts for 'bazel' and 'cppcheck'

+++ setup/install_bazel.sh
@@ -9,8 +9,10 @@ set -v
 # Install Bazel #
 #################
 
-sudo apt-get update
-sudo apt-get install -y openjdk-8-jdk curl
+if ! dpkg -s openjdk-8-jdk curl 1>/dev/null 2>&1; then
+    sudo apt-get update
+    sudo apt-get install -y openjdk-8-jdk curl
+fi
 
 echo "deb [arch=amd64] http://storage.googleapis.com/bazel-apt stable jdk1.8" |
     sudo tee /etc/apt/sources.list.d/bazel.list
